server/requirements.txt
bcrypt==4.3.0
click==8.2.1
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
flask-mysql-connector==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
